refactor(spf)!: move HLS engine and SpfMedia to src/playback-engines/hls/
The engine and its DOM adapter were buried under src/dom/playback-engine/,
which (a) conflated a cross-cutting composition (engine) with a DOM-bound
adapter (SpfMedia) under one directory, and (b) misnamed the subpath
(singular "playback-engine" won't age well once DASH/MoQ land, and "engine"
as a subpath category isn't where the adapter belongs).

New structure, anticipating multiple engines + possibly multiple HLS
compositions per engine:

  src/playback-engines/
    └── hls/
        ├── engine.ts      (was src/dom/playback-engine/hls-engine.ts)
        ├── adapter.ts     (was src/dom/playback-engine/adapter.ts)
        ├── index.ts
        ├── tests/
        └── tsconfig.json  — own subtree project, DOM lib

SpfMedia stays alongside the HLS engine for now because the current
adapter is coupled to this specific composition in practice (even though
structurally it just needs any composition with the right owners shape).
Revisit when a second engine lands and the engine-agnostic seam becomes
concrete.

BREAKING:
- @videojs/spf/playback-engine → @videojs/spf/hls (subpath renamed)
- @videojs/spf/dom no longer re-exports SpfMedia/SpfMediaMixin/
  spfMediaDefaultProps/SpfMediaAPI/SpfMediaProps — import them from
  @videojs/spf/hls instead.
- External consumers updated: packages/core/src/dom/media/simple-hls,
  packages/react/src/media/simple-hls-video, and the sandbox template
  for spf-segment-loading.

Scaffolding:
- Root tsconfig gains { path: "packages/spf/src/playback-engines/hls" }.
- packages/spf/package.json exports renamed.
- tsdown.config.ts entry "dom/playback-engine" → "hls".
- vitest.config.ts adds a "playback-engines" project (browser mode).

No behavior change. Build (160 files), 707 tests, typecheck, workspace
check, lint — all green.
