+++ packages/core/src/dom/media/simple-hls/index.ts
@@ -1,4 +1,4 @@
-import { SpfMediaMixin } from '@videojs/spf/dom';
+import { SpfMediaMixin } from '@videojs/spf/hls';
 import { HTMLVideoElementHost } from '../video-host';
 
 export class SimpleHlsMedia extends SpfMediaMixin(HTMLVideoElementHost) {}
