--- packages/spf/src/dom/playback-engine/index.ts
@@ -1,8 +1,0 @@
-export type { SpfMediaAPI } from './adapter';
-export { SpfMedia, SpfMediaMixin } from './adapter';
-export type {
-  HlsPlaybackEngineConfig,
-  HlsPlaybackEngineOwners,
-  HlsPlaybackEngineState,
-} from './hls-engine';
-export { createHlsPlaybackEngine } from './hls-engine';
