+++ packages/spf/src/playback-engines/hls/engine.ts
@@ -1,5 +1,17 @@
 import { type Composition, createComposition } from '../../core/composition/create-composition';
 import type { ReadonlySignal, Signal } from '../../core/signals/primitives';
+import type { TextTracksActor } from '../../dom/actors/text-tracks';
+import { endOfStream } from '../../dom/behaviors/end-of-stream';
+import { loadSegments } from '../../dom/behaviors/load-segments';
+import { provideTextTrackActors } from '../../dom/behaviors/provide-text-track-actors';
+import { setupMediaSource } from '../../dom/behaviors/setup-mediasource';
+import { setupSourceBuffers } from '../../dom/behaviors/setup-sourcebuffer';
+import { syncTextTracks } from '../../dom/behaviors/sync-text-tracks';
+import { trackCurrentTime } from '../../dom/behaviors/track-current-time';
+import { trackPlaybackInitiated } from '../../dom/behaviors/track-playback-initiated';
+import { updateDuration } from '../../dom/behaviors/update-duration';
+import type { SourceBufferActor } from '../../dom/media/source-buffer-actor';
+import { destroyVttResolver, resolveVttSegment } from '../../dom/text/resolve-vtt-segment';
 import type { BandwidthState } from '../../media/abr/bandwidth-estimator';
 import type { TextTrackSegmentLoaderActor } from '../../media/actors/text-track-segment-loader';
 import { calculatePresentationDuration } from '../../media/behaviors/calculate-presentation-duration';
@@ -9,18 +21,6 @@ import { resolvePresentation } from '../../media/behaviors/resolve-presentation'
 import { resolveTrack } from '../../media/behaviors/resolve-track';
 import { syncPreloadAttribute } from '../../media/behaviors/sync-preload-attribute';
 import { selectAudioTrack, selectTextTrack, selectVideoTrack } from '../../media/primitives/select-tracks';
-import type { TextTracksActor } from '../actors/text-tracks';
-import { endOfStream } from '../behaviors/end-of-stream';
-import { loadSegments } from '../behaviors/load-segments';
-import { provideTextTrackActors } from '../behaviors/provide-text-track-actors';
-import { setupMediaSource } from '../behaviors/setup-mediasource';
-import { setupSourceBuffers } from '../behaviors/setup-sourcebuffer';
-import { syncTextTracks } from '../behaviors/sync-text-tracks';
-import { trackCurrentTime } from '../behaviors/track-current-time';
-import { trackPlaybackInitiated } from '../behaviors/track-playback-initiated';
-import { updateDuration } from '../behaviors/update-duration';
-import type { SourceBufferActor } from '../media/source-buffer-actor';
-import { destroyVttResolver, resolveVttSegment } from '../text/resolve-vtt-segment';
 
 // ============================================================================
 // HLS Engine State & Owners
