+++ packages/spf/src/playback-engines/hls/tsconfig.json
@@ -1,0 +1,17 @@
+{
+  "extends": "../../../../../tsconfig.base.json",
+  "compilerOptions": {
+    "composite": true,
+    "lib": ["ES2022", "DOM", "DOM.Iterable"],
+    "exactOptionalPropertyTypes": false,
+    "declarationDir": "../../../types/playback-engines/hls"
+  },
+  "references": [
+    { "path": "../../../../utils" },
+    { "path": "../../core" },
+    { "path": "../../network" },
+    { "path": "../../media" },
+    { "path": "../../dom" }
+  ],
+  "include": ["./**/*.ts"]
+}
