+++ packages/spf/tsdown.config.ts
@@ -9,7 +9,7 @@ const createConfig = (mode: BuildMode): UserConfig => ({
   entry: {
     index: 'src/index.ts',
     dom: 'src/dom/index.ts',
-    'dom/playback-engine': 'src/dom/playback-engine/index.ts',
+    hls: 'src/playback-engines/hls/index.ts',
   },
   platform: 'neutral',
   format: 'es',
