+++ tsconfig.json
@@ -23,6 +23,7 @@
     { "path": "packages/spf/src/network" },
     { "path": "packages/spf/src/media" },
     { "path": "packages/spf/src/dom" },
+    { "path": "packages/spf/src/playback-engines/hls" },
 
     { "path": "packages/html" },
     { "path": "packages/react" },
